+++ src/Controllers/MapController.ts
@@ -596,11 +596,8 @@ module StreamStats.Controllers {
                                 if (item.layerName == "Streamgages" && k == "FeatureURL") {
                                     var siteNo = queryResult.properties[k].split('site_no=')[1];
                                     var SSgagepage = "vm.openGagePage('" + siteNo + "')";
-                                    var NWISpage = 'https://waterdata.usgs.gov/monitoring-location/' + siteNo;
-
-                                    var html = '<strong>Monitoring Location Page: </strong><a href="' + NWISpage + ' "target="_blank">link</a></br><strong>StreamStats Gage Page: </strong><a ng-click="' + SSgagepage + '">link</a></br>';
+                                    var html = '</br><br><button ng-click="' + SSgagepage + '" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button>';
                                     querylayers.append(html);
-
 
                                     this.angulartics.eventTrack('explorationTools', { category: 'Map', label: 'streamgageQuery' });
                                 }
@@ -1449,7 +1446,7 @@ module StreamStats.Controllers {
                         var gageButtonDiv = L.DomUtil.create('div', 'innerDiv');
 
                         gageButtonDiv.innerHTML = '<strong>Station ID: </strong>' + siteNo + '</br><strong>Station Name: </strong>' + feature.properties['Name'] + '</br><strong>Latitude: </strong>' + feature.geometry.coordinates[1] + '</br><strong>Longitude: </strong>' + feature.geometry.coordinates[0] + '</br><strong>Station Type</strong>: ' + feature.properties.StationType.name +
-                        '</br><strong>Monitoring Location Page: </strong><a href="' + NWISpage + ' "target="_blank">link</a></br><strong>StreamStats Gage Page: </strong><a id="gagePageLink" class="' + siteNo + '">link</a><br>';
+                        '</br><br><button id="gagePageLink" type="button" class="btn-blue fullwidth">Open StreamStats Gage Page</button>';
 
                         layer.bindPopup(gageButtonDiv);